+++ test/dummy/config/initializers/recording_studio_support.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RecordingStudioSupport.configure do |config|
+  # Hosts can set options here. Support pages do not need extra config in this slice.
+end